% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{tf_example_path}
\alias{tf_example_path}
\title{Path to a packaged example theory}
\usage{
tf_example_path(name)
}
\arguments{
\item{name}{File name, e.g. \code{"panic-network.theory.yaml"}. See
\code{\link[=tf_example_names]{tf_example_names()}}.}
}
\value{
The path to the installed file.
}
\description{
Path to a packaged example theory
}
\examples{
tf_read(tf_example_path("panic-network.theory.yaml"))$id
}
