% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{tf_jaccard}
\alias{tf_jaccard}
\title{Jaccard similarity of two token sets}
\usage{
tf_jaccard(a, b)
}
\arguments{
\item{a, b}{Character vectors of tokens (treated as sets).}
}
\value{
A numeric similarity in \code{[0, 1]}.
}
\description{
Returns 0.0 if both sets are empty, otherwise the size of the intersection
divided by the size of the union, rounded to 3 decimals.
}
\examples{
tf_jaccard(tf_tokens("arousal threat response"),
           tf_tokens("threat appraisal response"))
}
