% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{tf_example_names}
\alias{tf_example_names}
\title{Names of the packaged example theories}
\usage{
tf_example_names()
}
\value{
A sorted character vector of file names.
}
\description{
Names of the packaged example theories
}
\examples{
tf_example_names()
}
